Waited for network idle when creating and editing entities and in public forms
